A Japanese Translator
This is a work in progress. The program takes regular expression and
uses them to try and analyze the sentence structure of romanji.

For example, the code will search the sentence: "ashita, hon wo mimasu."
for key bits of linguistic information (such as verb stems and
particles) and organize them accordingly.
The translator will first scan the sentence for [ie]masu, and then
realize that anything it finds must be  a verb in the present polite
form, then [ie]mashita, etc.